Prune stage in reducer so properties can be reset
Previously null properties were stripped before sending update action - meaning that it was impossible to clear properties without doing an update with an overwrite.

This moves the prune call into the updateStage action, which is called after the new stage state has been calculated (applies to both merge and overwrite).