Create new EntitySelectField
This implements a new redux-form compatible field called
EntitySelectField. It replaces the NodeSelect and Edge select, which
did essentially the same thing, and had lots of complicated
duplication. It maintains the ability to create new node or edge types
inline, triggering the TypeEditor screen from a button.

It does away with the complex system of messaging between this screen,
and the consumer, deferring to the consuming component to determine
what to do when a new entity is created. Default behaviour is to
switch to the new entity if there are no existing entities.

Uses the redux-form onChange callback to reset form fields, and
provides a simple prop for user confirmation.

I replaced all existing instances of NodeSelect and EdgeSelect, as
well as updating the user of NativeSelect for selecting edges in
DyadCensusPrompts and SociogramPrompts.