Fix group variable preset creation doing nothing
Creating a categorical variable on the narrative interface preset
screen did not work. This is because creating a categorical variable
requires also creating multiple option values, and this functionality
was not connected to the narrative interface because it is intended
to be 'display only'.

Additionally, the previous variable select field could be configured
to disallow variable creation, but the new VariablePicker could not.

I implemented a `disallowCreation` prop for the variable picker that
restores this behaviour.