Refactored in WelcomeHeader.js
Moved the downloading process to userActions.js where openNetcanvas method is originally at, such that WelcomeHeader.js does not have to deal with complex logic in itself. At first I tried to move the downloading process to a saparate file in the utils folder, but somehow there are errors and I have not figured out how to solve them. It works fine in userActions.js so I will leave it there for now. I will go back to it later. At least it is no longer in WelcomeHeader.js anymore.