Users can switch between the preset and the custom modal
If users click on the create new node type button, the default modal will be the preset one so that users can select a new node from the preset list.
There is an underlined red text in the prompt so that users can switch to the custom modal, and switch back to the preset modal from a similar underlined red text.