Fix inability to remove field validation
Complex issue that revealed a series of unrelated bugs that have also
been fixed.

Root cause was that setting the validation field to `null` was causing
it to be removed. The `updateVariable` action was being called with
`merge: true`.

Basic fix was to fetch the existing variable definition (which required
refactoring several selectors), and handle the merging manually. This
allows us to completely remove any property that is set to `null`.

Further bugs/fixes:

- Additional investigation revealed that ego variables were not being
validated. Issued a fix in protocol-validation for this.
- This in turn revealed that entity-based filtering of validation
options was not being applied correctly (for example, `unique` is not
valid for `ego` variables). This was resolved by correctly passing
stage entity into the ValidationSection component.
- This in turn revealed that the TypeEditor contained a great deal of
obsolete code, including a VariableEditor component that was not being
used.