Release 6.4.3
This release contains a fix for a bug in the 6.4.2 release that caused
issues on the form interfaces. The bug was caused by errors in the logic
of the new `getAllVariablesByUUID()` function, which was incorrectly
expecting the root entity codebook property to have a `variables` key.

This was compounded by faulty logic in `makeGetVariable` selector,
which was incorrectly returning an empty object when the variable
definition was not found in the codebook. This caused the
`withFormHandlers` `changeField` handler to fail to correctly merge
existing variable properties.

Test coverage was added for all these scenarios.