Fixes not able to close stage editor when filter disabled
Redux form still validates fields that are not shown, even when their
value is null. We set fields to invisible with a value of null when they
are contained within a toggleable section, in order to "disable" them.

Redux form has no concept of disabling a field.

In this case, the hidden field representing the node panel filter was
still being validated, and the validation function was returning a
string indicating that the field was empty (because we set it to null in
order to reset it!). This error was hidden from the user, because the
field was hidden and deregistered.

The fix is to change the validation function to return undefined if the
value is null. Could also have been implemented as passing undefined as
the validation prop to the field based on if the toggleable section is
open or closed.