Add grid mixins, and fix z-index issue with datepicker and section editor
The newer versions of UI remove the grid mixins, which are unfortunately
widely used in Architect. I added them back in manually.

For the date picker z-index issue, this seems to have been caused
by switchiing the data picker to use an 'overlay', which was done in
2022 but not brought in to Architect because the UI version wasn't
correctly imported. Not sure why we didn't pick up the break when
we tested the PR for updating the BG color, as this should have
triggered the issue.

Regardless, the only way to fix this is to remove overflow: hidden
from the stage-editor-section, which allows the z-index of the
date picker's panel to take effect.

This alone would break the toggleable sections, since they use overflow
to hide the content when the section is being closed. I therefore
added the overflow state to the animation variants (where the open
variant has overflow: visible), and this seems to work.